Develop

+++ app/code/ForeverCompanies/CustomSales/Observer/Backend/Sales/OrderPlaceAfter.php
@@ -43,7 +43,7 @@ class OrderPlaceAfter implements ObserverInterface
         $statusString = stristr($refererUrl, 'status/');
         $statusString = str_replace('status/', '', $statusString);
         if ($statusString !== false) {
-            $position = strpos($statusString, "/");
+            $position = (int) strpos($statusString, "/");
             $order->setData('status', substr($statusString, 0, $position));
             $order->setData('state', substr($statusString, 0, $position));
             $order->setData('quote_expiration_date', date('Y-m-d', strtotime('+30 day')));
